Struct trillium_conn_id::ConnId

pub struct ConnId { /* private fields */ }

Trillium handler to set a identifier for every Conn.

By default, it will use an inbound x-request-id request header or if that is missing, populate a ten character random id. This handler will set an outbound x-request-id header as well by default. All of this behavior can be customized with ConnId::with_request_header, ConnId::with_response_header and ConnId::with_id_generator

Implementations

impl ConnId

pub fn new() -> Self

Constructs a new ConnId handler

let app = (ConnId::new().with_seed(1000), "ok"); // seeded for testing
assert_ok!(
    get("/").on(&app),
    "ok",
    "x-request-id" => "4fekClhof7"
);

assert_headers!(
    get("/")
        .with_request_header("x-request-id", "inbound")
        .on(&app),
    "x-request-id" => "inbound"
);

pub fn with_request_header( self, request_header: impl Into<HeaderName<'static>> ) -> Self

Specifies a request header to use. If this header is provided on the inbound request, the id will be used unmodified. To disable this behavior, see ConnId::without_request_header

let app = (
    ConnId::new().with_request_header("x-custom-id"),
    "ok"
);

assert_headers!(
    get("/")
        .with_request_header("x-custom-id", "inbound")
        .on(&app),
    "x-request-id" => "inbound"
);

pub fn without_request_header(self) -> Self

disables the default behavior of reusing an inbound header for the request id. If a ConnId is configured without_request_header, a new id will always be generated

pub fn with_response_header( self, response_header: impl Into<HeaderName<'static>> ) -> Self

Specifies a response header to set. To disable this behavior, see ConnId::without_response_header

let app = (
    ConnId::new()
        .with_seed(1000) // for testing
        .with_response_header("x-custom-header"),
    "ok"
);

assert_headers!(
    get("/").on(&app),
    "x-custom-header" => "4fekClhof7"
);

pub fn without_response_header(self) -> Self

Disables the default behavior of sending the conn id as a response header. A request id will be available within the application through use of ConnIdExt but will not be sent as part of the response.

pub fn with_id_generator<F>(self, id_generator: F) -> Self
where F: Fn() -> String + Send + Sync + 'static,

Provide an alternative generator function for ids. The default is a ten-character alphanumeric random sequence.

let app = (
    ConnId::new().with_id_generator(|| Uuid::new_v4().to_string()),
    "ok"
);

// assert that the id is a valid uuid, even if we can't assert a specific value
assert!(Uuid::parse_str(get("/").on(&app).headers_mut().get_str("x-request-id").unwrap()).is_ok());

pub fn with_seed(self, seed: u64) -> Self

seed a shared rng

this is primarily useful for tests