Trait trillium_conn_id::ConnIdExt

pub trait ConnIdExt {
    // Required method
    fn id(&self) -> &str;
}

Extension trait to retrieve an id generated by the ConnId handler

Required Methods

fn id(&self) -> &str

Retrieves the id for this conn. This method will panic if it is run before the ConnId handler.